+++ src/js/lib/Observer.js
@@ -1,0 +1,3 @@
+export default class Observer {
+  update() {}
+}
